+++ app/index.js
@@ -24,6 +24,7 @@ const {
 const fs = require("fs")
 const os = require("os")
 const path = require("path")
+const isSvg = require("is-svg")
 const {ElectronBlocker} = require("@cliqz/adblocker-electron")
 
 const version = process.env.npm_package_version || app.getVersion()
@@ -31,17 +32,16 @@ const printUsage = () => {
     console.log("Vieb: Vim Inspired Electron Browser\n")
     console.log("Usage: Vieb [options] <URLs>\n")
     console.log("Options:")
-    console.log(" --help         Print this help and exit")
-    console.log(" --version      Print program info with version and exit")
+    console.log(" --help             Print this help and exit")
+    console.log(" --version          Print program info with version and exit")
+    console.log(" --datafolder <dir> Store ALL Vieb data in this folder")
     console.log(
-        " --portable     Store ALL Vieb data in a relative ViebData folder")
+        " --erwic <file>     Open a fixed set of pages in a separate instance")
+    console.log("                    See 'Erwic.md' for usage and details")
     console.log(
-        " --erwic <file> Open a fixed set of pages in a separate instance")
-    console.log("                See 'Erwic.md' for usage and details")
+        " --debug            Open with Chromium and Electron debugging tools")
     console.log(
-        " --debug        Open with Chromium and Electron debugging tools")
-    console.log(
-        " --console      Open with the Vieb console (implied by --debug)")
+        " --console          Open with the Vieb console (implied by --debug)")
     console.log("\nAll arguments not starting with - will be opened as a url.")
     printLicense()
 }
@@ -138,15 +138,19 @@ if (args[0] === "app" || args[0] === app.getAppPath()) {
 const urls = []
 let enableDebugMode = false
 let showInternalConsole = false
-let portable = false
 let nextArgErwicConfig = false
+let nextArgDataFolder = false
 let erwic = null
+let datafolder = path.join(app.getPath("appData"), "Vieb")
 let customIcon = null
 args.forEach(arg => {
     arg = arg.trim()
     if (nextArgErwicConfig) {
         erwic = arg
         nextArgErwicConfig = false
+    } else if (nextArgDataFolder) {
+        datafolder = arg
+        nextArgDataFolder = false
     } else if (arg.startsWith("--")) {
         if (arg === "--help") {
             printUsage()
@@ -154,14 +158,14 @@ args.forEach(arg => {
         } else if (arg === "--version") {
             printVersion()
             app.exit(0)
-        } else if (arg === "--portable") {
-            portable = true
         } else if (arg === "--debug") {
             enableDebugMode = true
         } else if (arg === "--console") {
             showInternalConsole = true
         } else if (arg === "--erwic") {
             nextArgErwicConfig = true
+        } else if (arg === "--datafolder") {
+            nextArgDataFolder = true
         } else {
             console.log(`Unsupported argument: ${arg}`)
             printUsage()
@@ -172,14 +176,20 @@ args.forEach(arg => {
     }
 })
 if (nextArgErwicConfig) {
-    console.log(`The 'erwic' option requires an argument`)
+    console.log(`The 'erwic' option requires a file location`)
+    printUsage()
+    app.exit(1)
+}
+if (nextArgDataFolder) {
+    console.log(`The 'datafolder' option requires a directory location`)
     printUsage()
     app.exit(1)
 }
 app.setName("Vieb")
+datafolder = `${path.resolve(expandPath(datafolder.trim()))}/`
 if (erwic) {
-    if (portable) {
-        console.log("Portable mode can not be combined with an Erwic config")
+    if (!datafolder) {
+        console.log("Datafolder must be specified if using an Erwic config")
         printUsage()
         app.exit(1)
     }
@@ -204,45 +214,46 @@ if (erwic) {
         }
         customIcon = config.icon
     }
-    config.datafolder = config.datafolder && config.datafolder.trim
-        && path.resolve(expandPath(config.datafolder.trim()))
-    if (!config.datafolder) {
+    if (!datafolder) {
         console.log("Erwic config file requires a 'datafolder' location field")
         printUsage()
         app.exit(1)
     }
-    fs.mkdirSync(config.datafolder, {"recursive": true})
-    fs.writeFileSync(path.join(config.datafolder, "erwicmode"), "")
+    fs.writeFileSync(path.join(datafolder, "erwicmode"), "")
     if (!Array.isArray(config.apps)) {
-        console.log("Erwic config file requires a list of 'apps' ")
+        console.log("Erwic config file requires a list of 'apps'")
         printUsage()
         app.exit(1)
     }
-    config.apps
-        .filter(a => typeof a.name === "string" && typeof a.url === "string")
-        .forEach(a => {
-            if (typeof a.script === "string") {
-                a.script = expandPath(a.script)
-                if (a.script !== path.resolve(a.script)) {
-                    a.script = path.join(path.dirname(erwic), a.script)
-                }
-                if (!isFile(a.script)) {
-                    a.script = null
-                }
-            } else {
+    config.apps = config.apps.map(a => {
+        if (a.name && !a.container) {
+            // OLD remove fallback checks in 4.x.x
+            a.container = a.name
+        }
+        a.container = a.container?.replace(/[^A-Za-z0-9_]/g, "")
+        if (typeof a.script === "string") {
+            a.script = expandPath(a.script)
+            if (a.script !== path.resolve(a.script)) {
+                a.script = path.join(path.dirname(erwic), a.script)
+            }
+            if (!isFile(a.script)) {
                 a.script = null
             }
-        })
+        } else {
+            a.script = null
+        }
+        return a
+    }).filter(a => typeof a.container === "string" && typeof a.url === "string")
+    if (config.apps.length === 0) {
+        console.log("Erwic config file requires at least one app to be added")
+        console.log("Each app must have a 'container' name and a 'url'")
+        printUsage()
+        app.exit(1)
+    }
     urls.push(...config.apps)
-    app.setPath("appData", config.datafolder)
-    app.setPath("userData", config.datafolder)
-} else if (portable) {
-    app.setPath("appData", path.join(process.cwd(), "ViebData"))
-    app.setPath("userData", path.join(process.cwd(), "ViebData"))
-} else {
-    app.setPath("appData", path.join(app.getPath("appData"), "Vieb"))
-    app.setPath("userData", app.getPath("appData"))
 }
+app.setPath("appData", datafolder)
+app.setPath("userData", datafolder)
 if (showInternalConsole && enableDebugMode) {
     console.log("The --debug argument always opens the internal console")
 }
@@ -256,19 +267,30 @@ app.on("ready", () => {
     app.userAgentFallback = useragent()
     // Request single instance lock and quit if that fails
     if (app.requestSingleInstanceLock()) {
-        app.on("second-instance", (_, commandLine) => {
+        app.on("second-instance", (_, newArgs) => {
             if (mainWindow.isMinimized()) {
                 mainWindow.restore()
             }
             mainWindow.focus()
-            commandLine = commandLine.slice(1)
-            if (commandLine[0] === "app") {
-                commandLine = commandLine.slice(1)
+            newArgs = newArgs.slice(1)
+            if (newArgs[0] === "app" || newArgs[0] === app.getAppPath()) {
+                newArgs = newArgs.slice(1)
             }
-            mainWindow.webContents.send("urls",
-                commandLine.filter(arg => !arg.startsWith("-")))
+            const newUrls = []
+            let ignoreNextArg = false
+            newArgs.forEach(arg => {
+                if (arg === "--erwic" || arg === "--datafolder") {
+                    ignoreNextArg = true
+                } else if (ignoreNextArg) {
+                    ignoreNextArg = false
+                } else if (!arg.startsWith("-")) {
+                    newUrls.push(arg)
+                }
+            })
+            mainWindow.webContents.send("urls", newUrls)
         })
     } else {
+        console.log(`Sending urls to existing instance in ${datafolder}`)
         app.exit(0)
     }
     // Init mainWindow
@@ -401,7 +423,6 @@ app.on("login", (e, contents, _, auth, callback) => {
         return
     }
     if (loginAttempts.includes(contents.id)) {
-        contents.stop()
         loginAttempts.splice(loginAttempts.indexOf(contents.id), 1)
         return
     }
@@ -701,10 +722,12 @@ const writeDownloadsToFile = () => {
 }
 const permissionHandler = (_, permission, callback, details) => {
     if (permission === "media") {
-        if (details.mediaTypes && details.mediaTypes.includes("video")) {
+        if (details.mediaTypes?.includes("video")) {
             permission = "camera"
-        } else {
+        } else if (details.mediaTypes?.includes("audio")) {
             permission = "microphone"
+        } else {
+            permission = "mediadevices"
         }
     }
     const permissionName = `permission${permission.toLowerCase()}`
@@ -837,7 +860,12 @@ ipcMain.on("download-favicon", (_, fav, location, webId, linkId, url) => {
     request.on("response", res => {
         const data = []
         res.on("end", () => {
-            fs.writeFileSync(location, Buffer.concat(data))
+            const file = Buffer.concat(data)
+            if (isSvg(file)) {
+                location += ".svg"
+                fav += ".svg"
+            }
+            fs.writeFileSync(location, file)
             mainWindow.webContents.send("favicon-downloaded", linkId, url, fav)
         })
         res.on("data", chunk => {
@@ -946,6 +974,13 @@ ipcMain.on("hide-window", () => {
         mainWindow.hide()
     }
 })
+ipcMain.on("add-devtools", (_, pageId, devtoolsId) => {
+    const page = webContents.fromId(pageId)
+    const devtools = webContents.fromId(devtoolsId)
+    page.setDevToolsWebContents(devtools)
+    page.openDevTools()
+    devtools.executeJavaScript("window.location.reload()")
+})
 ipcMain.on("destroy-window", () => {
     cancellAllDownloads()
     mainWindow.destroy()
@@ -965,10 +1000,9 @@ ipcMain.on("set-window-title", (_, title) => {
 })
 ipcMain.handle("show-message-dialog", (_, options) => dialog.showMessageBox(
     mainWindow, options))
-ipcMain.handle("list-cookies",
-    () => session.fromPartition("persist:main").cookies.get({}))
-ipcMain.handle("remove-cookie", (_, url, name) => session.fromPartition(
-    "persist:main").cookies.remove(url, name))
+ipcMain.handle("list-cookies", e => e.sender.session.cookies.get({}))
+ipcMain.handle("remove-cookie",
+    (e, url, name) => e.sender.session.cookies.remove(url, name))
 // Operations below are sync
 ipcMain.on("override-global-useragent", (e, globalUseragent) => {
     app.userAgentFallback = globalUseragent || useragent()