+++ app/js/command.js
@@ -163,7 +163,7 @@ const quitall = () => {
         HISTORY.writeHistToFile(true)
     }
     TABS.saveTabs()
-    UTIL.clearContainerTabs()
+    UTIL.clearTempContainers()
     if (SETTINGS.get("cache") !== "full") {
         UTIL.clearCache()
     }
@@ -177,22 +177,49 @@ const quitall = () => {
     ipcRenderer.send("destroy-window")
 }
 
+const openDevTools = (position = null, trailingArgs = false) => {
+    if (trailingArgs) {
+        UTIL.notify("The devtools command takes a single optional argument",
+            "warn")
+        return
+    }
+    if (!position) {
+        position = SETTINGS.get("devtoolsposition")
+    }
+    if (position === "window") {
+        TABS.currentPage().openDevTools()
+    } else if (position === "tab") {
+        TABS.addTab({"devtools": true})
+    } else if (position === "vsplit") {
+        TABS.addTab({
+            "switchTo": false,
+            "devtools": true,
+            "callback": id => {
+                PAGELAYOUT.add(id, "hor", !SETTINGS.get("splitright"))
+                TABS.switchToTab(tabIndexById(id))
+            }
+        })
+    } else if (position === "split") {
+        TABS.addTab({
+            "switchTo": false,
+            "devtools": true,
+            "callback": id => {
+                PAGELAYOUT.add(id, "ver", !SETTINGS.get("splitbelow"))
+                TABS.switchToTab(tabIndexById(id))
+            }
+        })
+    } else {
+        UTIL.notify("Invalid devtools position specified, must be one of: "
+            + "window, vsplit, split or tab", "warn")
+    }
+}
+
 const openSpecialPage = (specialPage, section = null) => {
-    // Switch to already open special page if available
-    let alreadyOpen = false
-    TABS.listTabs().forEach((tab, index) => {
-        // The list of tabs is ordered, the list of pages isn't
-        const page = TABS.tabOrPageMatching(tab)
-        if (UTIL.pathToSpecialPageName(page.src).name === specialPage) {
-            alreadyOpen = true
-            TABS.switchToTab(index)
-        }
-    })
     // Open the url in the current or new tab, depending on currently open page
     const pageUrl = UTIL.specialPagePath(specialPage, section)
     const isNewtab = UTIL.pathToSpecialPageName(
         TABS.currentPage().src).name === "newtab"
-    if (TABS.currentPage().src === "" || alreadyOpen || isNewtab) {
+    if (!TABS.currentPage().isLoading() && isNewtab) {
         TABS.navigateTo(pageUrl)
     } else {
         TABS.addTab({"url": pageUrl})
@@ -353,6 +380,7 @@ const addSplit = (method, leftOrAbove, args) => {
     if (args.length === 0) {
         TABS.addTab({
             "switchTo": false,
+            "container": SETTINGS.get("containersplitpage"),
             "callback": id => {
                 PAGELAYOUT.add(id, method, leftOrAbove)
                 TABS.switchToTab(tabIndexById(id))
@@ -371,6 +399,7 @@ const addSplit = (method, leftOrAbove, args) => {
     } else {
         TABS.addTab({
             "url": UTIL.stringToUrl(args.join(" ")),
+            "container": SETTINGS.get("containersplitpage"),
             "switchTo": false,
             "callback": id => {
                 PAGELAYOUT.add(id, method, leftOrAbove)
@@ -450,7 +479,7 @@ const commands = {
     "quit": quit,
     "qa": quitall,
     "quitall": quitall,
-    "devtools": () => TABS.currentPage().openDevTools(),
+    "devtools": openDevTools,
     "reload": reload,
     "v": () => openSpecialPage("version"),
     "version": () => openSpecialPage("version"),
@@ -491,7 +520,6 @@ const noArgumentComands = [
     "quit",
     "qa",
     "quitall",
-    "devtools",
     "reload",
     "v",
     "version",