+++ CHANGELOG.md
@@ -13,14 +13,62 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 ## Unreleased
 
+### Added
+
+- Local file suggestions to explore mode and to relevant commands
+- New 'suggestfiles' setting to control when/if file suggestions should appear ('all' by default)
+- Setting to control if file suggestions should appear before/after history in explore mode: suggestfilesfirst
+- Red border in command mode when the entered command is invalid (for example due to unmatched quotes)
+- Setting "closablepinnedtabs" which is disabled by default (meaning pinned tabs can't be closed)
+- Setting "containernewtab" to control which container all new tabs should use (default is still "main")
+- Setting "containersplitpage" to control which container all split pages should use (default is "s:usecurrent")
+- Setting "containerstartuppage" to control which container the startup pages should use (CLI arguments, default is still "main")
+- Setting "containercolors" to show container tabs in a custom color based on the container name
+- Setting "containershowname" to optionally show the name of the container between the mode and url in the navbar
+- Setting "containerkeeponreopen" to toggle the remembering the container name and to use the "containernewtab" setting instead
+- Special container names to open tabs externally, in tabs with a matching domain, or use the same container as the current tab
+- Support for SVG favicons by giving it an explicit ".svg" extension if detected by "is-svg"
+- Devtools that can be opened as a split window or in a separate tab, in addition to the existing windowed developer tools
+
+### Changed
+
+- Suggestions are now scrollable when they don't fit the window
+- The setting 'suggesthistory' has been renamed 'suggestexplore' as it might also include file suggestions now
+- Command suggestions are now by default set to 1000, as the operation is simple and suggestions are scrollable
+- Increased the upper limit for the number of suggestions (for both command and explore mode)
+- Pinned tabs can no longer be closed, unless the 'closablepinnedtabs' setting is enabled
+- Search setting parsing, you can now include %s as a substitute for the entered search
+- Default duckduckgo search engine options now include a persistent header and infinite scrolling
+- Tabs are now always modifiable in Erwic mode, because containers can now be managed at runtime
+- Multiple tabs of the same special page can now be opened using the commands, as some of them are container specific
+- The startup argument "--portable" has been replaced with "--datafolder", use "--datafolder ./ViebData" for old functionality
+- Erwic now uses the "datafolder" startup argument instead of a JSON field for the datafolder location configuration
+
+### Fixed
+
+- Typo in the "storenewvisits" setting (there will be no automatic migration from the name with a typo to the correct one)
+- Pointer not updating the location when brining back the GUI while in fullscreen
+- Permission for media devices sometimes being detected as a microphone permission
+
+### Deprecated
+
+- Old tabs file format (it's now a list of objects, instead of a list of strings
+- Conversion from old format will be removed in future Vieb 4.x.x releases
+- Supplying a "name" to an Erwic app, use a "container" field instead
+
+### Removed
+
+- Old line-based history format parser (1.x.x users should update to 2.x.x first, if they want to keep their history)
+
 ### Fixed
 
 - Abort error for aborting page loads in the debug console (using --debug or --console)
+- Don't keep closed pages in memory while the 'keeprecentlyclosed' setting is off
 
 ### Security
 
-- Electron 10.1.2 (was 10.1.1)
-- Chromium 85.0.4183.98 (was 85.0.4183.93)
+- Electron 11.0.0-beta.11 (was 10.1.1)
+- Chromium 87.0.4251.1 (was 85.0.4183.93)
 
 ## [2.4.0](https://github.com/Jelmerro/Vieb/compare/2.3.0...2.4.0) - 2020-09-05
 
@@ -602,7 +650,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 ### Changed
 
 - Moved most information regarding configuration from the readme to the help page
-- Changed the default search engine from duckduckgo to duckduckgo with a dark theme
+- Default duckduckgo search engine to use a dark theme
 - Improved CLI startup arguments for packaged apps
 - Split the preload into multiple separate preloads
 - Improved electron builder configuration (separate file with much better settings)